Lots of little changes!
I've got a load of inserters for weird unicode symbols...
...commands to send things flush right and flush left...

But I'm commiting now to cut my losses.

I really need to figure out a way to fix this habit of committing loads
of non-related stuff because it's too much of a pain to change things.

One solution would be to put things in different files... that'd
seperate changes!