Fix new-alias.
I've removed the non-interactive version, because it didn't make sense
to have one.

The interactive version is now fixed, but you need to be careful when
using it so you don't accidentally get rid of necessary quotes.

For example:

 $ new-alias "ls='ls --color=auto'"
 Description: coloured ls output

will result in

 alias ls='ls --color=auto'

being placed at the bottom of lib/aliases.